pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyper-audio"
version = "0.1.0"
description = "AI-powered audio processing pipeline for podcast voice replacement"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "James", email = "james@example.com"},
]

dependencies = [
    # Core ML frameworks
    "torch>=2.0.0",
    "transformers>=4.20.0",
    "huggingface-hub>=0.15.0",
    
    # Audio processing
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pydub>=0.25.0",
    "scipy>=1.9.0",
    
    # AI models
    "openai-whisper>=20230314",
    "speechbrain>=0.5.0",
    "demucs>=4.0.0",
    
    # Speaker diarization
    "pyannote.audio>=3.0.0",
    
    # Utilities
    "numpy>=1.21.0",
    "tqdm>=4.64.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']